+++ Nginx/var/www/kibana/html/administration.php
@@ -19,9 +19,13 @@ if(isset($_GET['update'])) {
 exec("chmod 777 /var/www/kibana/html/update/update -R");
 exec("rm /var/www/kibana/html/update/updateStatus.log");
 $TAG=$_GET["update"];
-exec("sed -i '3s/.*$/$TAG=\"'$TAG'\"/g' /home/amadmin/box4s/BOX4s-main/update.sh");
+$TAG=escapeshellcmd($TAG);
+# Man könnte hier TAG in das Script schreiben,
+# allerdings nicht nötig, wir bekommen per GitLab API ohnehin die neueste Version raus.
+# Daher ist der Parameter hier nicht weiter verwendet
 //Update.sh muss per Installscript und UpdateScript www-data gehören und +x bekommen
-exec('sudo /home/amadmin/box4s/BOX4s-main/update.sh >/dev/null &2>/dev/null &');
+// $updatePid holds the pid of update.sh
+exec('sudo /home/amadmin/box4s/BOX4s-main/update.sh >/dev/null 2>&1 & echo $!', $updatePid);
 //exec("chmod 777 /var/www/kibana/html/update/ -R");
 }
 ?>
@@ -35,99 +39,138 @@ function sleep(milliseconds) {
 
 async function openUpdateModal() {
 $('.ui.modal').modal('show');
-var updateStatus="";
+var consoleEcho="";
+var updateRunning=true;
 do {
 	// Iframe dynamicContent.php wird alle 2 sec neu aufgerufen und der Status geupdatet
 	var iframe = document.getElementById('dynamic');
-	updateStatus=iframe.contentWindow.consoleEcho;
 	await sleep(1000);
-	iframe.src="dynamicContent.php";
+  <?php if (isset($updatePid)): ?>
+    iframe.src="/dynamicContent.php?pid="+<?=$updatePid[0]?>;
+  <?php else: ?>
+    iframe.src="/dynamicContent.php"
+  <?php endif; ?>
+  consoleEcho=iframe.contentWindow.consoleEcho;
+  updateRunning=iframe.contentWindow.updateRunning;
 	await sleep(1000);
-	console.log(updateStatus);
-	document.getElementById("updateStatus").innerHTML = updateStatus;
-} while (updateStatus.match("Update abgeschlossen") != "Update abgeschlossen");
+	document.getElementById("consoleEcho").innerHTML = consoleEcho;
+} while (updateRunning);
 // Nachdem Update abgeschlossen im Status erscheit wird die Seite freigegeben und die Elemente entsprechend mit neuem Content versehen
-document.getElementById("statusUpdateMessage").innerHTML = "Das Update ist abgeschlossen";
-document.getElementById("updateHeader").innerHTML = "Sie können das Fenster jetzt schließen";
+document.getElementById("statusUpdateMessage").innerHTML = "Das Update ist abgeschlossen.";
+document.getElementById("statusUpdateMessageLong").innerHTML = "Sie können das Fenster jetzt schließen.";
 document.getElementById("dimmer").setAttribute('class','ui inverted dimmer');
+$('#circle').attr('class','large icon green checkmark');
 
 }
 </script>
 
 <?php
-// Logik für die Anzeige der bereits installierten und installierbaren Versionen. CurVer ist die aktuell installierte Version.
+// Get TAG Count
 exec("curl -s https://gitlab.am-gmbh.de/api/v4/projects/it-security%2Fb4s/repository/tags --header 'PRIVATE-TOKEN: p3a72xCJnChRkMCdUCD6' | python3 -c 'import sys, json; print(len(json.load(sys.stdin)))'",$tagCount);
-exec("tail /home/amadmin/box4s/BOX4s-main/VERSION",$curVer);
+// exec("curl -s https://lockedbox-bugtracker.am-gmbh.de/api/v4/projects/AM-GmbH%2Fbox4s/repository/tags --header 'PRIVATE-TOKEN: Lmp3tZkURptSjWsn7tyC' | python3 -c 'import sys, json; print(len(json.load(sys.stdin)))'",$tagCount);
+
+// CurVer ist die aktuell installierte Version.
+exec("tail /home/amadmin/VERSION",$curVer);
+$pos = strpos($curVer[0], 'v');
+if ($pos === 0) {
+  $curVer[0] = substr_replace($curVer[0], '', $pos, strlen('v'));
+}
 for($ctr=0;$ctr<$tagCount[0];$ctr++){
+  // get TAG Names
 	exec('curl -s https://gitlab.am-gmbh.de/api/v4/projects/it-security%2Fb4s/repository/tags --header "PRIVATE-TOKEN: p3a72xCJnChRkMCdUCD6" | python3 -c "import sys, json; print(json.load(sys.stdin)['.$ctr.'][\'name\'])"',$tags[$ctr]);
+  // exec('curl -s https://lockedbox-bugtracker.am-gmbh.de/api/v4/projects/AM-GmbH%2Fbox4s/repository/tags --header "PRIVATE-TOKEN: Lmp3tZkURptSjWsn7tyC" | python3 -c "import sys, json; print(json.load(sys.stdin)['.$ctr.'][\'name\'])"',$tags[$ctr]);
+  $pos = strpos($tags[$ctr][0], 'v');
+  if ($pos === 0) {
+    $tags[$ctr] = substr_replace($tags[$ctr], '', $pos, strlen('v'));
+  }
 }
 ?>
 </head>
 
 
-<?php
-if(isset($_GET['update'])) { echo('<body onload="openUpdateModal()">');}
-else{ echo("<body>"); }
-?>
+<?php if (isset($_GET['update'])): ?>
+  <body onload="openUpdateModal()">
+<?php else: ?>
+  <body>
+<?php endif; ?>
 <iframe src="dynamicContent.php" name="dynamic" width="0px" height="0px" id="dynamic"> </iframe>
 
 <div class="ui modal">
-  <div id="updateHeader" class="header">Bitte lassen Sie das Fenster geöffnet</div>
+  <div class="header">Update</div>
   <div class="content">
     <p id="updateContent"></p>
-    <p id="statusUpdateMessage">Der Updatevorgang läuft</p>
+    <div class="ui icon message">
+      <i id="circle" class="notched circle loading icon"></i>
+      <div class="content">
+        <div id="statusUpdateMessage" class="header">
+          Der Updatevorgang läuft..
+        </div>
+        <p id="statusUpdateMessageLong">Das System wird aktualisiert. Bitte lassen Sie das Fenster geöffnet.</p>
+      </div>
+    </div>
    <div class="ui segment">
   <div class="ui active inverted dimmer" id="dimmer">
-    <div class="ui text loader">Loading</div>
   </div>
-  <p id="updateStatus"><br> </p>
+  <p id="consoleEcho"><br> </p>
 </div></div></div>
 
-<div class="ui divider"></div>
-<div class="ui grid padded  ">
+<div class="ui grid padded">
 <div class="two wide column">
 <p> Verfügbare Updates </p>
 <div class="ui container">
 <table class="ui celled structured table">
   <thead>
     <tr>
-      <th with="50px">Version</th>
+      <th>Version</th>
       <th>Update</th>
     </tr>
 </thead>
 <tbody>
-
 <?php
-for($ctr=0;$ctr<$tagCount[0];$ctr++) {
-echo('<tr>');
-echo('<td class="right aligned">');
-echo($tags[$ctr][0]);
-echo('</td><td>');
-if ($ctr<$tagCount[0]){
-if ($curVer[0]==$tags[$ctr+1][0]){
-	echo('<form method="get" action="administration.php">');
-
-	echo('<input type="hidden" value="'.$tags[$ctr][0].'" name="update">');
-	echo('<button class="ui button" type="submit">Update</button>');
-	echo('</form>');
-}elseif ($curVer[0]==$tags[$ctr][0]){
-	echo('<i class="large green checkmark icon"></i>');
-}elseif ($curVer[0]> $tags[$ctr][0]){
-	echo('<i class="large green checkmark icon"></i>');
-}elseif ($curVer[0]< $tags[$ctr][0]){
-echo('<i class="large red close icon"></i>');
-}
-echo ('</td>');
-}
+$ctr = 0;
+?>
+<tr>
+  <td class="right aligned"><?= $tags[$ctr][0] ?></td>
+  <td>
+<?php // https://www.php.net/manual/de/function.version-compare.php ?>
+<?php if (version_compare($curVer[0], $tags[$ctr][0], "<")): ?>
+    <form style="margin:0;" class="ui form" method="get" action="administration.php">
+      <input type="hidden" value="<?= $tags[$ctr][0] ?>" name="update">
+      <button class="ui button" type="submit">Update</button>
+    </form>
+<?php else: ?>
+  <i class="large green checkmark icon"></i>
+<?php endif; ?>
+</td>
+</tr>
+<?php
+if(++$ctr < $tagCount) {
+  // Rest of the tags if more available
+for(;$ctr<$tagCount[0];$ctr++) {
+?>
+<tr>
+<td class="right aligned">
+  <?php if (version_compare($curVer[0], $tags[$ctr][0], "==")): ?>
+    <b><?= $tags[$ctr][0] ?></b>
+  <?php else: ?>
+    <?= $tags[$ctr][0] ?>
+  <?php endif; ?>
+</td>
+<td>
+<?php if (version_compare($curVer[0], $tags[$ctr][0], ">=")): ?>
+  <i class="large green checkmark icon"></i>
+<?php else: ?>
+  <i class="large angle double up icon"></i>
+<?php endif; ?>
+</td>
+</tr>
+<?php
+  }
 }
 ?>
-
-</tr>
-  </tbody>
+</tbody>
 </table>
 </div>
-</div>
-<div class="four wide column">
 </div>
 </div>
 </body>